network: move broadcast_transaction move network to interface.py
